requirements.txt
python-telegram-bot>=20.0
nest_asyncio>=1.5.6
python-dotenv>=1.0.0
pytz>=2023.3
google-genai>=0.1.0
pandas